move listener to own mod
