+++ ssdp-rs/src/lib.rs
@@ -8,9 +8,13 @@
 
 use std::net::{IpAddr, Ipv4Addr, SocketAddr};
 
+pub mod listener;
 pub mod message;
 pub mod search;
 pub mod udp;
+
+pub use listener::Listener;
+pub use search::Searcher;
 
 pub const MULTICAST_IP: Ipv4Addr = Ipv4Addr::new(239, 255, 255, 250);
 pub const SSDP_PORT: u16 = 1900;
